// src/app/carritos/interfaces/compra.interface.ts
export interface CompraDTO {
    pagoId:          number | null;
    clienteId:       number | null;
    nombreCliente:   string | null;
    apellidoCliente: string | null;
    emailCliente:    string | null;
    direccionId:     number | null;
    calle:           string | null;
    altura:          string | null;
    piso:            string | null;
    departamento:    string | null;
    localidad:       string | null;
    provincia:       string | null;
}