// src/app/compra/interfaces/cliente.interface.ts
export interface Cliente {
    id?:                number;
    fechaCreacion?:     Date;
    fechaModificacion?: null;
    fechaBaja?:         null;
    nombre?:            string;
    apellido?:          string;
    dni?:               string;
    email?:             string;
    password?:          string;
    direccion?:         Direccion;
    ventas?:            CarritoElement[];
    carritos?:          CarritoElement[];
}

export interface CarritoElement {
    id?:                number;
    fechaCreacion?:     Date | null;
    fechaModificacion?: null;
    fechaBaja?:         null;
    direccion?:         Direccion;
    pago?:              Pago;
    carrito?:           CarritoCarrito;
}

export interface CarritoCarrito {
    id?:                number;
    fechaCreacion?:     Date;
    fechaModificacion?: null;
    fechaBaja?:         null;
    precioTotal?:       number;
    items?:             Item[];
}
// ...
export interface Pago {
    id?:                number;
    fechaCreacion?:     Date;
    fechaModificacion?: null;
    fechaBaja?:         null;
    nombre?:            PagoNombre;
    tipo?:              Tipo;
}
// ...
export interface Direccion {
    id?:                number;
    fechaCreacion?:     Date;
    fechaModificacion?: null;
    fechaBaja?:         null;
    calle?:             Calle;
    altura?:            string;
    piso?:              string;
    depto?:             string;
    localidad?:         Localidad;
    provincia?:         Provincia;
}

export enum Calle {
    Bonifacini = "Bonifacini",
    Campos = "campos",
}

export enum Localidad {
    SANMartin = "San martin",
}

export enum Provincia {
    BuenosAires = "Buenos Aires",
}